signal-detection/requirements-test.txt
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27.0
psycopg2-binary>=2.9.0
